backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
llama-index>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-qdrant>=0.9.0
qdrant-client>=1.7.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.5
rank-bm25>=0.2.2
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.12.0
httpx>=0.26.0
requests>=2.31.0
numpy>=1.26.0
pymongo>=4.6.0
motor>=3.3.0
